pyproject.toml
[build-system]

requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "peak-mas"
version = "1.0.11"
description = "Python-based framework for heterogeneous agent communities"
readme = "README.md"
authors = [{ name = "Bruno Ribeiro", email = "brgri@isep.ipp.pt" }, {name = "Luis Gomes", email = "log@isep.ipp.pt"}]
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Environment :: Console",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
keywords = ["framework", "multiagent", "agent-based", "ecosystem", "spade", "xmpp"]
dependencies = [
    "aiohttp_cors",
    "aioxmpp",
    "pandas",
    "spade == 3.2.2",
    "PyYAML"
]
requires-python = "==3.9.6"

[project.optional-dependencies]
build = ["build", "twine"]
dev = ["black", "isort", "mypy", "bumpver", "pipreqs", "autoflake", "pylint"]

[project.urls]
Homepage = "https://www.gecad.isep.ipp.pt/peak"
Github = "https://github.com/gecad-group/peak-mas"

[project.scripts]
peak = "peak.__main__:main"

[tool.bumpver]
current_version = "1.0.11"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "chore: bump version {old_version} -> {new_version}"
commit          = true
tag             = true
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"src/peak/__init__.py" = ["{version}"]

[tool.isort]
profile                   = "black"